Simplified samples. Refactor tests.
