+++ store/src/test/java/com/dropbox/android/external/store4/testutil/FlowSubjectTest.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.dropbox.android.external.store4.util
+package com.dropbox.android.external.store4.testutil
 
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
