+++ store/src/test/java/com/dropbox/android/external/store4/testutil/SimplePersisterAsFlowable.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.dropbox.android.external.store4.util
+package com.dropbox.android.external.store4.testutil
 
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.BroadcastChannel
@@ -130,3 +130,9 @@ internal class KeyTracker<Key> {
         }
     }
 }
+
+@ExperimentalCoroutinesApi
+suspend fun <Key, Output> InMemoryPersister<Key, Output>.asFlowable() = SimplePersisterAsFlowable(
+    reader = this::read,
+    writer = this::write
+)
